Multiple fixes
-- LUIE Print to Chat function now uses string.format, this fixes wierd
problems, and also fixes problems with Social Indicators. This is cool,
we can list this as a compatibility feature. There's literally no reason
it needs to be string.format either.
-- Changed menu option for throttles to MS instead of seconds. Gives
more customization. Considering adding the same function for gold, but
generally gold is acquired in one big loot chunk so its not really
needed.
-- Fixed Experience throttle to use the option rather than always being
a flat 5s delay. Not sure how I missed this one.
-- Fixed issue with AP throttle
-- Active Non-Combat Pets buff icon now automatically hidden in Cyrodiil
where pets are disabled
-- Fixed potential overlapping on PLAYER frame, as well as fixed target
NPC frames to not clip the name as much