Nonfunctional update push
Please note there's a lot of cobbled together functions in this mess
currently - This is not intended to be a functional build.
-- Note that I am currently working on re-coding a significant portion
of Chat Announcements to be much cleaner and more efficient - as well as
adding a few more customization options.
-- Expect more passes in the future on Buff/Debuff icons, etc
-- Combat Cloud will either be updated or integrated at some point in
the future (before I do any major release).

Update Notes:
-- Stripped most functionality from CombatInfo
-- Removed DamageMeter and CombatLog components entirely

-- Buffs now display a countdown border based on the total duration of
the buff rather than not starting until 8 seconds remaining.

-- Updates for CA Currency, new formatting options (Not integrated with
Mail, Trade, Loot, etc yet)
-- Disabled some throttling functions in LOOT temporarily until I decide
what to do with indexing
-- Removed some zo_callLaters as part of an efficiency cleanup
-- Lorebooks/Achievements Announcements now prehook Zenimax
CenterScreenAnnounce events, THIS IS STILL WIP - eventually all
components with an alert or CSA component will be switched to this
method for improved efficiency and control
-- Significant stripped and updated Experience function. Removed current
XP counter because the amount of nonsense it took to keep this value
accurate was ridiculous. Much cleaner function will make future work far
easier.
-- Minor updates to various components

OTHER:
-- bindings.xml cleared of bindings but not removed, will keep this here
if I have anything bound in the future.
-- Strings are a bit unorganized, and a lot of them need to be removed.
Will be working on cleaning this up over time.