SCB Update
-- Bar highlight and proc functions slightly changed, previously
switching bars after an active highlight on the other bar had faded
would result in the backdrop/label flashing for a moment before being
removed, this no longer occurs.
-- Removed the return statement at the end of the function in
EVENT_EFFECT_CHANGED for fake ground targeted abilities. This means we
can set ground target fake auras to work on any ability outgoing with
the player as a source. First use of this was for the Monster sets that
summon an NPC to also create their timer duration effect on the player.
-- Updated Monster sets to use new Fake Ground auras
-- Added missing monste sets - Sentinel of Rkugamz + Slimecraw