CA Update
-- Updated friend/ignore messages to new format
-- Added quest share messages to new format
-- When loot is pickpocketed the appropriate context message now appears
again.
-- Edited hooked group invite function (TryGroupInviteByName) to ignore
the true/false conditional to display message or not. This allows
invites sent from the Group list to display chat announcements and
alerts without having to hook that function as well.
-- Added alternate strings to message sent from player radial menu for
invites. (Replaces "name" with [name] link since this type of message is
always sent with a char/account name we know)