Fix doubled guild join announcement
With the release of Elsweyr, guilds received the ability to see a list of currently invited players within the guild roster. LUI's code was set to check for when a guild member was added to trigger the join notification, but this event also triggers upon invite because the player is added to the roster at that time. This fix performs an additional check to make sure the event is a true join and not an invite.